Separated the Cluster and Delegate codes
