+++ src/app/clusters/electrical-power-measurement-server/app_config_dependent_sources.cmake
@@ -17,5 +17,7 @@ TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
     "${CLUSTER_DIR}/ElectricalPowerMeasurementCluster.cpp"
+    "${CLUSTER_DIR}/ElectricalPowerMeasurementCluster.h"
+    "${CLUSTER_DIR}/ElectricalPowerMeasurementDelegate.h"
     "${CLUSTER_DIR}/electrical-power-measurement-server.h"
 )
