+++ src/app/clusters/electrical-power-measurement-server/app_config_dependent_sources.gni
@@ -13,5 +13,7 @@
 # limitations under the License.
 app_config_dependent_sources = [
   "ElectricalPowerMeasurementCluster.cpp",
+  "ElectricalPowerMeasurementCluster.h",
+  "ElectricalPowerMeasurementDelegate.h",
   "electrical-power-measurement-server.h",
 ]
