+++ src/app/clusters/electrical-power-measurement-server/electrical-power-measurement-server.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include "ElectricalPowerMeasurementCluster.h"
+#include "ElectricalPowerMeasurementDelegate.h"
 #include <app/AttributeAccessInterface.h>
 
 namespace chip {
